Start splitting chain module up
