+++ populus/chain/__init__.py
@@ -1,0 +1,23 @@
+from populus.contracts.exceptions import (  # noqa: F401
+    NoKnownAddress,
+    UnknownContract,
+    BytecodeMismatchError,
+)
+
+from .geth import (  # noqa: F401
+    BaseGethChain,
+    LocalGethChain,
+    MainnetChain,
+    TemporaryGethChain,
+    TestnetChain,
+    reset_chain,
+)
+from .external import (  # noqa: F401
+    ExternalChain,
+)
+from .tester import (  # noqa: F401
+    TesterChain,
+)
+from .testrpc import (  # noqa: F401
+    TestRPCChain,
+)
