+++ populus/chain/base.py
@@ -1,59 +1,10 @@
 from __future__ import absolute_import
 import itertools
-import copy
-
-try:
-    from contextlib import ExitStack
-except ImportError:
-    from contextlib2 import ExitStack
 
 from pylru import lrucache
 
-from geth import (
-    DevGethProcess,
-    LiveGethProcess,
-    TestnetGethProcess,
-    LoggingMixin,
-)
-
-from web3 import (
-    IPCProvider,
-    HTTPProvider,
-)
-
 from populus.legacy.contracts import (
     get_contract_library_dependencies,
-)
-
-from populus.utils.functional import (
-    cached_property,
-)
-from populus.utils.networking import (
-    get_open_port,
-    wait_for_connection,
-)
-from populus.utils.wait import (
-    Wait,
-)
-from populus.utils.filesystem import (
-    remove_file_if_exists,
-    remove_dir_if_exists,
-    get_blockchains_dir,
-    tempdir,
-)
-from populus.utils.linking import (
-    link_bytecode_by_name,
-)
-from populus.utils.contracts import (
-    construct_contract_factories,
-    package_contracts,
-)
-from populus.utils.chains import (
-    get_chaindata_dir,
-    get_dapp_dir,
-    get_nodekey_path,
-    get_geth_ipc_path,
-    get_geth_logfile_path,
 )
 
 from populus.migrations.migration import (
@@ -62,101 +13,28 @@ from populus.migrations.migration import (
 from populus.migrations.registrar import (
     get_registrar,
 )
+from populus.contracts.exceptions import (
+    NoKnownAddress,
+    BytecodeMismatchError,
+    UnknownContract,
+)
+
+from populus.utils.contracts import (
+    construct_contract_factories,
+    package_contracts,
+)
+from populus.utils.functional import (
+    cached_property,
+)
+from populus.utils.linking import (
+    link_bytecode_by_name,
+)
+from populus.utils.wait import (
+    Wait,
+)
 
 
-TESTNET_BLOCK_1_HASH = '0xad47413137a753b2061ad9b484bf7b0fc061f654b951b562218e9f66505be6ce'
-MAINNET_BLOCK_1_HASH = '0x88e96d4537bea4d9c05d12549907b32561d3bf31f45aae734cdc119f13406cb6'
-
-
-def reset_chain(data_dir):
-    chaindata_dir = get_chaindata_dir(data_dir)
-    remove_dir_if_exists(chaindata_dir)
-
-    dapp_dir = get_dapp_dir(data_dir)
-    remove_dir_if_exists(dapp_dir)
-
-    nodekey_path = get_nodekey_path(data_dir)
-    remove_file_if_exists(nodekey_path)
-
-    geth_ipc_path = get_geth_ipc_path(data_dir)
-    remove_file_if_exists(geth_ipc_path)
-
-
-class LoggedDevGethProcess(LoggingMixin, DevGethProcess):
-    def __init__(self, project_dir, blockchains_dir, chain_name, overrides):
-        stdout_logfile_path = get_geth_logfile_path(
-            project_dir,
-            chain_name,
-            'stdout'
-        )
-        stderr_logfile_path = get_geth_logfile_path(
-            project_dir,
-            chain_name,
-            'stderr',
-        )
-        super(LoggedDevGethProcess, self).__init__(
-            overrides=overrides,
-            chain_name=chain_name,
-            base_dir=blockchains_dir,
-            stdout_logfile_path=stdout_logfile_path,
-            stderr_logfile_path=stderr_logfile_path,
-        )
-
-
-class LoggedTestnetGethProccess(LoggingMixin, TestnetGethProcess):
-    def __init__(self, project_dir, geth_kwargs):
-        super(LoggedTestnetGethProccess, self).__init__(
-            geth_kwargs=geth_kwargs,
-        )
-
-
-class LoggedMainnetGethProcess(LoggingMixin, LiveGethProcess):
-    def __init__(self, project_dir, geth_kwargs):
-        super(LoggedMainnetGethProcess, self).__init__(
-            geth_kwargs=geth_kwargs,
-            stdout_logfile_path=get_geth_logfile_path(
-                project_dir,
-                'mainnet',
-                'stdout'
-            ),
-            stderr_logfile_path=get_geth_logfile_path(
-                project_dir,
-                'mainnet',
-                'stderr',
-            ),
-        )
-
-
-class ContractError(Exception):
-    """
-    Base Exception class for errors raised by the Chain contract API.
-    """
-    pass
-
-
-class BytecodeMismatchError(ContractError):
-    """
-    Indicates there is a bytecode mismatch.
-    """
-    pass
-
-
-class NoKnownAddress(ContractError):
-    """
-    Raised when the address for a requested contract is not known.
-    """
-    pass
-
-
-class UnknownContract(ContractError):
-    """
-    Raised when the requested contract name is not found in the compiled
-    contracts.
-    """
-    pass
-
-
-class Chain(object):
+class BaseChain(object):
     """
     Base class for how populus interacts with the blockchain.
 
@@ -449,29 +327,6 @@ class Chain(object):
         return package_contracts(contract_classes)
 
 
-class ExternalChain(Chain):
-    """
-    Chain class to represent an externally running blockchain that is not
-    locally managed.  This class only houses a pre-configured web3 instance.
-    """
-    @property
-    def has_registrar(self):
-        return 'registrar' in self.config
-
-    @cached_property
-    def registrar(self):
-        if not self.has_registrar:
-            raise KeyError(
-                "The configuration for the {0} chain does not include a "
-                "registrar.  Please set this value to the address of the "
-                "deployed registrar contract.".format(self.chain_name)
-            )
-        return get_registrar(
-            self.web3,
-            address=self.config['registrar'],
-        )
-
-
 def testrpc_fn_proxy(fn_name):
     def inner(self, *args, **kwargs):
         fn = getattr(self.rpc_methods, fn_name)
@@ -480,7 +335,7 @@ def testrpc_fn_proxy(fn_name):
     return inner
 
 
-class BaseTesterChain(Chain):
+class BaseTesterChain(BaseChain):
     provider = None
 
     has_registrar = True
@@ -560,228 +415,4 @@ class BaseTesterChain(Chain):
             link_dependencies=link_dependencies,
             *args,
             **kwargs
-        )
-
-
-class TestRPCChain(BaseTesterChain):
-    port = None
-
-    def get_web3_config(self):
-        base_config = super(TestRPCChain, self).get_web3_config()
-        config = copy.deepcopy(base_config)
-        config['provider.settings.port'] = self.port
-        return config
-
-    def __enter__(self):
-        if self._running:
-            raise ValueError("The TesterChain is already running")
-
-        if self.port is None:
-            self.port = get_open_port()
-
-        self._running = True
-
-        self.provider = self.web3.currentProvider
-        self.rpc_methods = self.provider.server.application.rpc_methods
-
-        self.rpc_methods.full_reset()
-        self.rpc_methods.rpc_configure('eth_mining', False)
-        self.rpc_methods.rpc_configure('eth_protocolVersion', '0x3f')
-        self.rpc_methods.rpc_configure('net_version', 1)
-        self.rpc_methods.evm_mine()
-
-        wait_for_connection('127.0.0.1', self.port)
-        return self
-
-    def __exit__(self, *exc_info):
-        if not self._running:
-            raise ValueError("The TesterChain is not running")
-        try:
-            self.provider.server.stop()
-            self.provider.server.close()
-            self.provider.thread.kill()
-        except AttributeError:
-            self.provider.server.shutdown()
-            self.provider.server.server_close()
-        finally:
-            self._running = False
-
-
-class TesterChain(BaseTesterChain):
-    def __enter__(self):
-        if self._running:
-            raise ValueError("The TesterChain is already running")
-
-        self._running = True
-
-        self.provider = self.web3.currentProvider
-        self.rpc_methods = self.provider.rpc_methods
-
-        self.rpc_methods.full_reset()
-        self.rpc_methods.rpc_configure('eth_mining', False)
-        self.rpc_methods.rpc_configure('eth_protocolVersion', '0x3f')
-        self.rpc_methods.rpc_configure('net_version', 1)
-        self.rpc_methods.evm_mine()
-
-        return self
-
-    def __exit__(self, *exc_info):
-        if not self._running:
-            raise ValueError("The TesterChain is not running")
-        self._running = False
-
-
-GETH_KWARGS = {
-    'data_dir',
-    'geth_executable',
-    'max_peers',
-    'network_id',
-    'no_discover',
-    'mine',
-    'autodag',
-    'miner_threads',
-    'nice',
-    'unlock',
-    'password',
-    'port',
-    'verbosity',
-    'ipc_disable',
-    'ipc_path',
-    'ipc_api',
-    'ws_enabled',
-    'ws_enabled',
-    'ws_addr',
-    'ws_origins',
-    'ws_port',
-    'ws_api',
-    'rpc_enabled',
-    'rpc_addr',
-    'rpc_port',
-    'rpc_api',
-    'prefix_cmd',
-    'suffix_args',
-    'suffix_kwargs',
-}
-
-
-class BaseGethChain(Chain):
-    stack = None
-    geth = None
-
-    def initialize_chain(self):
-        # context manager shenanigans
-        self.stack = ExitStack()
-        self.geth = self.get_geth_process_instance()
-
-    @property
-    def geth_kwargs(self):
-        return self.config.get('chain.settings', {})
-
-    @property
-    def has_registrar(self):
-        return 'registrar' in self.config
-
-    @cached_property
-    def registrar(self):
-        if not self.has_registrar:
-            raise KeyError(
-                "The configuration for the {0} chain does not include a "
-                "registrar.  Please set this value to the address of the "
-                "deployed registrar contract.".format(self.chain_name)
-            )
-        return get_registrar(
-            self.web3,
-            address=self.config['registrar'],
-        )
-
-    def get_geth_process_instance(self):
-        raise NotImplementedError("Must be implemented by subclasses")
-
-    def __enter__(self, *args, **kwargs):
-        self.stack.enter_context(self.geth)
-
-        if self.geth.is_mining:
-            self.geth.wait_for_dag(600)
-        if self.geth.ipc_enabled:
-            self.geth.wait_for_ipc(60)
-        if self.geth.rpc_enabled:
-            self.geth.wait_for_rpc(60)
-
-        self._running = True
-
-        return self
-
-    def __exit__(self, *exc_info):
-        self.stack.close()
-        self._running = False
-
-
-class LocalGethChain(BaseGethChain):
-    def get_geth_process_instance(self):
-        return LoggedDevGethProcess(
-            project_dir=self.project.project_dir,
-            blockchains_dir=self.project.blockchains_dir,
-            chain_name=self.chain_name,
-            overrides=self.geth_kwargs,
-        )
-
-    def get_web3_config(self):
-        base_config = super(LocalGethChain, self).get_web3_config()
-        config = copy.deepcopy(base_config)
-        if issubclass(base_config.provider_class, IPCProvider):
-            config['provider.settings.ipc_path'] = self.geth.ipc_path
-        elif issubclass(base_config.provider_class, HTTPProvider):
-            config['provider.settings.endpoint_uri'] = "http://127.0.0.1:{0}".format(
-                self.geth.rpc_port,
-            )
-        else:
-            raise ValueError("Unknown provider type")
-        return config
-
-
-class TemporaryGethChain(BaseGethChain):
-    def get_geth_process_instance(self):
-        tmp_project_dir = self.stack.enter_context(tempdir())
-        blockchains_dir = get_blockchains_dir(tmp_project_dir)
-
-        return LoggedDevGethProcess(
-            project_dir=self.project.project_dir,
-            blockchains_dir=blockchains_dir,
-            chain_name=self.chain_name,
-            overrides=self.geth_kwargs,
-        )
-
-    def get_web3_config(self):
-        base_config = super(TemporaryGethChain, self).get_web3_config()
-        config = copy.deepcopy(base_config)
-        config['provider.settings.ipc_path'] = self.geth.ipc_path
-        return config
-
-    has_registrar = True
-    _registrar = None
-
-    @property
-    def registrar(self):
-        if self._registrar is None:
-            RegistrarFactory = get_registrar(self.web3)
-            deploy_txn_hash = RegistrarFactory.deploy()
-            registrar_address = self.wait.for_contract_address(deploy_txn_hash)
-            self._registrar = RegistrarFactory(address=registrar_address)
-
-        return self._registrar
-
-
-class TestnetChain(BaseGethChain):
-    def get_geth_process_instance(self):
-        return LoggedTestnetGethProccess(
-            project_dir=self.project.project_dir,
-            geth_kwargs=self.geth_kwargs,
-        )
-
-
-class MainnetChain(BaseGethChain):
-    def get_geth_process_instance(self):
-        return LoggedMainnetGethProcess(
-            project_dir=self.project.project_dir,
-            geth_kwargs=self.geth_kwargs,
         )
