+++ tests/chain-management/test_chain_get_contract_factory.py
@@ -1,5 +1,6 @@
-import os
 import pytest
+
+import os
 
 from populus.utils.filesystem import (
     remove_file_if_exists,
@@ -24,6 +25,7 @@ from populus.migrations.writer import (
 from populus import Project
 
 
+@pytest.mark.slow
 def test_get_contract_factory_with_no_dependencies(temp_chain):
     chain = temp_chain
 
@@ -34,6 +36,7 @@ def test_get_contract_factory_with_no_dependencies(temp_chain):
     assert Math.bytecode_runtime == MATH['bytecode_runtime']
 
 
+@pytest.mark.slow
 def test_get_contract_factory_with_missing_dependency(temp_chain):
     chain = temp_chain
 
@@ -42,6 +45,7 @@ def test_get_contract_factory_with_missing_dependency(temp_chain):
 
 
 
+@pytest.mark.slow
 def test_get_contract_factory_with_declared_dependency(temp_chain):
     chain = temp_chain
 
@@ -66,6 +70,7 @@ def test_get_contract_factory_with_declared_dependency(temp_chain):
     assert Multiply13.bytecode_runtime == expected_runtime
 
 
+@pytest.mark.slow
 def test_get_contract_factory_with_registrar_dependency(temp_chain,
                                                         library_13):
     chain = temp_chain
@@ -93,6 +98,7 @@ def test_get_contract_factory_with_registrar_dependency(temp_chain,
     assert Multiply13.bytecode_runtime == expected_runtime
 
 
+@pytest.mark.slow
 def test_with_bytecode_mismatch_in_registrar_dependency(temp_chain,
                                                         library_13):
     chain = temp_chain
@@ -110,6 +116,7 @@ def test_with_bytecode_mismatch_in_registrar_dependency(temp_chain,
         chain.get_contract_factory('Multiply13')
 
 
+@pytest.mark.slow
 def test_bytecode_comes_from_project_if_no_migrations(temp_chain):
     project = Project()
     assert not project.migrations
